Rewrite expander algorithm
We now are more deterministic: for each pass we collect the packages
to install and install them in groups. Thus, the ordering of the
packages can no longer influence the result. There is a slight chance
that this may lead to an expansion error for a build that worked
before.

We also improved the expander interface in order to move the
expansion code to a separate file.